web/requirements.txt
django
psycopg2-binary
python-dotenv
djangorestframework_simplejwt
Pillow
drf-yasg
django-filter
gunicorn
